:sparkles: :fire: :truck: :recycle: :art: :bug: :white_check_mark: Added youtube API
- Refactored code
- Moved mocked files to a new folder
- Mocked API calls
- Created tests for custom hooks
