--- .env
@@ -1,1 +1,0 @@
-YOUTUBE_API_KEY=AIzaSyB-h_01yfbjyBX-IRZPyoytjfzx1jQ2_HU
