+++ public/index.html
@@ -10,12 +10,7 @@
 
     <script src="https://apis.google.com/js/api.js"></script>
     <script>
-      function start() {
-        gapi.client.init({
-          apiKey: 'AIzaSyCJwdVPNkyBkUry0R6FWUPUX-Bi6qeE0C0',
-        });
-      }
-      gapi.load('client', start);
+      gapi.load('client');
     </script>
     <!--
       manifest.json provides metadata used when your web app is installed on a
