+++ src/components/App/App.component.jsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import styled, { ThemeProvider } from 'styled-components';
 import { LayoutWrapper } from '../Layout';
-import { SearchWrapper } from '../../wrappers/SearchWrapper';
 import HomeVideos from '../HomeVideos/HomeVideos';
 import { lightTheme, darkTheme } from '../../providers/themes';
 
@@ -16,21 +15,19 @@ function App() {
     setIsLightTheme(!isLightTheme);
   };
 
-  const getTheme = () => ({
+  const getThemeConfig = () => ({
     theme: isLightTheme ? lightTheme : darkTheme,
     switchTheme,
   });
 
   return (
-    <ThemeProvider theme={getTheme()}>
-      <SearchWrapper>
-        <LayoutWrapper>
-          <StyledWelcome>
-            <h1>Welcome to the challenge!</h1>
-          </StyledWelcome>
-          <HomeVideos />
-        </LayoutWrapper>
-      </SearchWrapper>
+    <ThemeProvider theme={getThemeConfig()}>
+      <LayoutWrapper>
+        <StyledWelcome>
+          <h1>Welcome to the challenge!</h1>
+        </StyledWelcome>
+        <HomeVideos />
+      </LayoutWrapper>
     </ThemeProvider>
   );
 }
