+++ src/components/ui/Card/Card.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import 'jest-styled-components';
 import { getByRole } from '@testing-library/react';
-import { renderWithTheme } from '../../../utils/testing';
+import { renderWithTheme } from '../../../utils';
 import { lightTheme } from '../../../providers/themes';
 import Card from './Card';
