--- src/hooks/useYoutubeAPI.js
@@ -1,15 +1,0 @@
-// import { useEffect, useState } from "react";
-
-// export const useYoutubeAPI = () => {
-//   /*global gapi*/
-//   /*eslint no-undef: "error"*/
-//   const [client, setClient] = useState(null);
-
-//   const start = () => {
-//     setClient(gapi.client);
-//   }
-
-//   useEffect(() => { gapi.load('client', start); });
-
-//   return {client};
-// }
